+++ src/main/java/io/github/syst3ms/skriptparser/registration/SkriptRegistration.java
@@ -24,10 +24,10 @@ import io.github.syst3ms.skriptparser.pattern.PatternElement;
 import io.github.syst3ms.skriptparser.pattern.PatternParser;
 import io.github.syst3ms.skriptparser.pattern.RegexGroup;
 import io.github.syst3ms.skriptparser.pattern.TextElement;
-import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue;
-import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue.State;
-import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue.Usage;
-import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValues;
+import io.github.syst3ms.skriptparser.registration.context.ContextValue;
+import io.github.syst3ms.skriptparser.registration.context.ContextValue.State;
+import io.github.syst3ms.skriptparser.registration.context.ContextValue.Usage;
+import io.github.syst3ms.skriptparser.registration.context.ContextValues;
 import io.github.syst3ms.skriptparser.registration.tags.Tag;
 import io.github.syst3ms.skriptparser.registration.tags.TagInfo;
 import io.github.syst3ms.skriptparser.registration.tags.TagManager;
