+++ src/main/java/io/github/syst3ms/skriptparser/registration/context/package-info.java
@@ -1,4 +1,4 @@
 @ParametersAreNonnullByDefault
-package io.github.syst3ms.skriptparser.registration.contextvalues;
+package io.github.syst3ms.skriptparser.registration.context;
 
 import javax.annotation.ParametersAreNonnullByDefault;
